+++ products/static/products/js/products.js
@@ -12,6 +12,7 @@ document.addEventListener('DOMContentLoaded', async (e) => {
             offCanvas.show();
 
             activateQuantityChange();
+            activateAddToCartButton(() => offCanvas.hide());
         });
     });
 
@@ -20,49 +21,22 @@ document.addEventListener('DOMContentLoaded', async (e) => {
     })
 });
 
-const activateQuantityChange = () => {
-    document.querySelectorAll('.btn-quantity').forEach(btn => {
-        btn.addEventListener('click', (e) => {
-            e.preventDefault();
-            let type = btn.getAttribute('data-type');
-            changeQtyAddToCart(type);
-        });
+function activateAddToCartButton(callback) {
+    document.querySelector('form.add-to-cart').addEventListener('submit', async (e) => {
+        e.preventDefault();
+        let form = e.target;
+        let url = form.getAttribute('action');
+        let method = form.getAttribute('method');
+        let data = new FormData(form);
+    
+        await fetch(url, {
+            method: method,
+            body: data
+        }).then(res => res.json());
+        
+        callback();
     });
-    document.querySelector('input.input-number').addEventListener('change', (e) => {
-        let input = document.querySelector('input.input-number');
-
-        let value = parseInt(e.target.value);
-        let max = e.target.getAttribute('max');
-        let min = e.target.getAttribute('min');
-        if (value > max) {
-            input.value = max;
-        } else if (value < min) {
-            input.value = min;
-        }
-        changeQtyAddToCart('custom');
-    });
-};
-
-const changeQtyAddToCart = (type) => {
-    let input = document.querySelector('input.input-number');
-    let quantity = parseInt(input.value);
-    if (type === 'plus') {
-        if (quantity < input.getAttribute('max')) {
-            input.value = quantity + 1;
-        }
-    } else if (type === 'minus') {
-        if (quantity > 1) {
-            input.value = quantity - 1;
-        }
-    }
-    calculateTotal(parseFloat(input.value));
-};
-
-const calculateTotal = (quantity) => {
-    let total = document.querySelector('.add-total');
-    let price = parseFloat(document.querySelector('.offcanvas-body .price').innerText.split('$')[1]);
-    total.innerText = (price * quantity).toFixed(2);
-};
+}
 
 // Filter by allergy
 function filterProducts(productList, allergy) {
