+++ products/templates/products/products.html
@@ -96,5 +96,6 @@
 
 {% block extra_js %}
 <script src="{% static 'products/js/categoryNav.js' %}"></script>
+<script src="{% static 'products/js/quantityInput.js' %}"></script>
 <script src="{% static 'products/js/products.js' %}"></script>
 {% endblock %}
